requirements.txt
docker-py==1.6.0
pyaml==15.8.2
PyYAML==3.11
requests==2.8.1
six==1.10.0
websocket-client==0.34.0
Werkzeug==0.11.10
